Move the config stuff before the index.js file
The index.js file loads stuff asynchronously, and not
until the window loads, so the config.js file probably
will load in time every time. But the index.js does
fire stuff of that depends on some variables in the config
file, so figured I'd move it above just in case.